Refactor Matrix.tsx to extract telemetry logic to useMatrixTelemetry.ts
