+++ src/components/effects/Matrix/Matrix.tsx
@@ -2,13 +2,7 @@ import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 
 import { cn } from "@/utils/commonUtils";
 import { HackTerminal } from "./HackTerminal";
-import {
-  buildHackCompleteConsole,
-  DEFAULT_CONSOLE_PROMPT,
-  HACK_COMPLETE_FEEDBACK,
-  HACKER_TYPER_CORPUS,
-  type HackCompleteConsoleParams,
-} from "./hackCopy";
+import { DEFAULT_CONSOLE_PROMPT, HACKER_TYPER_CORPUS } from "./hackCopy";
 import { ATTEMPT_START_PROGRESS } from "./hackTuning";
 import "./matrix.scss";
 import { getMatrixRainIntensity } from "./matrixRainIntensity";
@@ -18,6 +12,7 @@ import { useHackAttempt } from "./useHackAttempt";
 import { useHackInteraction } from "./useHackInteraction";
 import { useHackProgressDecay } from "./useHackProgressDecay";
 import { useMatrixRain } from "./useMatrixRain";
+import { useMatrixTelemetry } from "./useMatrixTelemetry";
 
 interface MatrixProps {
   isVisible: boolean;
@@ -45,21 +40,6 @@ const Matrix = ({ isVisible, onDismiss, onMatrixReady }: MatrixProps) => {
 
   useMatrixRain(canvasRef, isVisible, rainIntensityRef);
   const completionTriggeredRef = useRef(false);
-  const [attemptStart] = useState(() => Date.now());
-  const [matrixCoordinate] = useState<string>(() => {
-    const sector = Math.floor(Math.random() * 64)
-      .toString(16)
-      .toUpperCase()
-      .padStart(2, "0");
-    const node = Math.floor(Math.random() * 4096)
-      .toString(16)
-      .toUpperCase()
-      .padStart(3, "0");
-    return `${sector}:${node}`;
-  });
-  const [signalSeed] = useState<number>(
-    () => Math.floor(Math.random() * 900) + 100,
-  );
   const { completeHack, showHackCompleteFeedback } = useUnlock();
   const easterEggTriggeredRef = useRef<boolean>(false);
   const [easterEggs, setEasterEggs] = useState<number[]>([]);
@@ -67,7 +47,6 @@ const Matrix = ({ isVisible, onDismiss, onMatrixReady }: MatrixProps) => {
     () => Array.from({ length: 24 }, () => HACKER_TYPER_CORPUS).join("\n"),
     [],
   );
-  const completionTelemetryRef = useRef<HackCompleteConsoleParams | null>(null);
 
   const {
     hackInputRef,
@@ -112,6 +91,12 @@ const Matrix = ({ isVisible, onDismiss, onMatrixReady }: MatrixProps) => {
     triggerIdleFailure,
   });
 
+  const { completionTelemetry } = useMatrixTelemetry({
+    isHackComplete,
+    setHackFeedback,
+    setHackingBuffer,
+  });
+
   const handleDismissEasterEgg = useCallback(
     (eggId: number) => {
       setEasterEggs((prev) => prev.filter((id) => id !== eggId));
@@ -149,7 +134,6 @@ const Matrix = ({ isVisible, onDismiss, onMatrixReady }: MatrixProps) => {
   );
 
   const consoleDisplay = hackingBuffer || DEFAULT_CONSOLE_PROMPT;
-  const completionTelemetry = completionTelemetryRef.current;
   const showConsoleCursor = !isHackComplete;
 
   const handleContainerClick = useCallback(
@@ -205,62 +189,6 @@ const Matrix = ({ isVisible, onDismiss, onMatrixReady }: MatrixProps) => {
       resetHackStream();
     }
   }, [hackingBuffer, resetHackStream]);
-
-  useEffect(() => {
-    if (!isHackComplete) {
-      completionTelemetryRef.current = null;
-      return;
-    }
-
-    if (!completionTelemetryRef.current) {
-      const now = Date.now();
-      const elapsedSeconds = Math.max(
-        0,
-        Math.round((now - attemptStart) / 1000),
-      );
-
-      const runtimeDisplay = new Date(elapsedSeconds * 1000)
-        .toISOString()
-        .substring(11, 19);
-      const timecodeDisplay = new Date(now).toISOString().substring(11, 19);
-
-      const oscillation = Math.sin(elapsedSeconds / 2) * 4;
-      const progressBonus = 100 / 3;
-      const signalGain = Math.round(
-        signalSeed / 10 + oscillation + progressBonus,
-      );
-
-      const base = Math.floor(signalSeed / 3);
-      const jitter = (elapsedSeconds % 7) * 3;
-      const signalChannel = (base + jitter).toString().padStart(3, "0");
-
-      completionTelemetryRef.current = {
-        matrixCoordinate,
-        runtimeDisplay,
-        timecodeDisplay,
-        signalGain,
-        signalChannel,
-      };
-    }
-
-    const completeReadout = buildHackCompleteConsole({
-      matrixCoordinate: completionTelemetryRef.current.matrixCoordinate,
-      runtimeDisplay: completionTelemetryRef.current.runtimeDisplay,
-      timecodeDisplay: completionTelemetryRef.current.timecodeDisplay,
-      signalGain: completionTelemetryRef.current.signalGain,
-      signalChannel: completionTelemetryRef.current.signalChannel,
-    });
-
-    setHackFeedback(HACK_COMPLETE_FEEDBACK);
-    setHackingBuffer(completeReadout);
-  }, [
-    isHackComplete,
-    matrixCoordinate,
-    setHackFeedback,
-    setHackingBuffer,
-    attemptStart,
-    signalSeed,
-  ]);
 
   useEffect(() => {
     if (!isHackComplete || completionTriggeredRef.current) {
